\name{AffymetrixCdfFile}
\docType{class}
\alias{AffymetrixCdfFile}
\alias{AffymetrixCdfFile-class}
\title{Placeholder For AffymetrixCdfFile Documentation}
\description{
  The documentation is available by typing ?aroma.affymetrix::AffymetrixCdfFile,
  but to avoid a check warning in the \code{Repitools} package, this help file
  is present.
}
