\name{AffymetrixCelSet}
\docType{class}
\alias{AffymetrixCelSet}
\alias{AffymetrixCelSet-class}
\title{Placeholder For AffymetrixCelSet Documentation}
\description{
  The documentation is available by typing ?aroma.affymetrix::AffymetrixCelSet,
  but to avoid a check warning in the \code{Repitools} package, this help file
  is present.
}
