\name{expr}
    \docType{data}
    \alias{expr}
    \title{Vector of expression differences}
    \description{
        The t-statistics of differences in expression for genes on chromosome 21
        between prostate cancer and normal epithelial cells.
    }
    \usage{expr}
    \format{A numeric matrix, 309 rows and 1 column.}
    \keyword{datasets}
