\name{hcRegions}
     \docType{data}
     \alias{hcRegions}
     \title{Masking files for hg19}
     \description{
	 File to mask out areas of the genome that are prone to causing false positives in ChIP-seq and other sequencing based functional assays,
     as proposed by Pickrell et al. (2011), Bioinformatics 27: 2144-2146, http://eqtl.uchicago.edu/Home.html. 
     }
     \usage{hcRegions}
     \format{A GRanges object created using the bedfile provided on http://eqtl.uchicago.edu/Masking/seq.cov1.ONHG19.bed.gz.}
     \source{Pickrell et al. (2011), Bioinformatics 27: 2144-2146.}
     \keyword{datasets}

