\name{MappabilitySource}
\docType{class}
\alias{MappabilitySource}
\alias{MappabilitySource-class}

\title{Superclass for datatypes that can refer to genome mappability data.}

\description{
  This class is simply the union of \code{character} and \code{\linkS4class{BSgenome}}
  classes.
}

\author{Dario Strbenac}
