\name{samplesList}
    \docType{data}
    \alias{samplesList}
    \title{Short Reads from Cancer and Normal}
    \description{
        Short reads that mapped to chromosome 21 in an Illumina sequencing
        experiment that was looking for differences between healthy epithelial
        and prostate cancer cells. The DNA was immunoprecipitated by a DNA
        methylation binding antibody.
    }
    \usage{samplesList}
    \format{A \code{\linkS4class{GRangesList}}.}
    \keyword{datasets}

